+++ dotnet/Cache/src/EC.MemoryCache/RELEASE-NOTES.md
@@ -1,12 +1,38 @@
-### 1.0.1
+## 2.0.0
 
 **Date:** 2024-02-20
 
-Support .NET8
+### What's new
+
+Support .NET 9
+
+Add APIs:  
++ IEnumerable<object> GetAllKeys  
++ IReadOnlyList<object> ScanKeys  
 
 
-### 0.1.1
+### Breaking changes
+
+Remove APIs:
+ + IDictionary<object, DateTimeOffset?> GetAllKeys  
+ + IDictionary<object, DateTimeOffset?> GetKeys
+
+
+
+## 1.0.1
+
+**Date:** 2024-02-20
+
+### What's new
+
+Support .NET 8
+
+
+
+## 0.1.1
 
 **Date:** 2023-04-11  
+
+### What's new
 
 Rename assembly to avoid conflict